src/ap2/pyproject.toml
[project]
name = "pokemon-marketplace-ap2"
version = "1.0.0"
description = "AP2 Protocol implementation for Pokemon Marketplace"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.9.0",
    "httpx>=0.27.0",
    "mcp>=1.3.0",
    "google-genai>=0.7.1",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.8.0",
    "cryptography>=41.0.0",
    "sqlalchemy>=2.0.44",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["../../tests/ap2"]
pythonpath = ["."]
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "ap2: AP2 protocol tests",
]
